long.max_value 변경

+++ algorithm_java/twopointer/BOJ_2230.java
@@ -36,7 +36,7 @@ public class BOJ_2230 {
 		Arrays.sort(arr);
 		
 		int en=0;
-		long result = Integer.MAX_VALUE;
+		long result = Long.MAX_VALUE;
 		
 		for (int i = 0; i < N; i++) {
 			while (en < N && arr[en]-arr[i] < M ) {
